refactor the whole app
