+++ internal/book/service/mocks/mockService.go
