+++ internal/book/service/service.go
